separate terraform generate and copilot generate
